Time: 3 ms (99.72%), Space: 5.9 MB (31.85%) - LeetHub

+++ 9-palindrome-number/9-palindrome-number.cpp
@@ -3,7 +3,7 @@ public:
     bool isPalindrome(int x) {
     
         int iDigit = 0,iNo1 = 0;
-        long long int iNo2 = 0;
+        long int iNo2 = 0;
         
         if(x < 0)
         {
